Move block information into block package
